+++ src/components/LoginSection.tsx
@@ -1,12 +1,10 @@
 import React from 'react'
 import { useSelector, useDispatch } from 'react-redux';
-import { RootState } from '../reducers/index';
-import {LoginHandler, LogoutHandler} from '../actions/actions';
+import { RootState } from '../redux/modules/reducer';
+import {LoginHandler, LogoutHandler} from '../redux/modules/account';
 
 const LoginSection: React.FC = () => {
-  //로그인 하지 않았을때는 로그인 버튼
-  //로그인 했을 때는 마이페이지 버튼
-  const loginState = useSelector((state: RootState) => state.LoginReducer)
+  const loginState = useSelector((state: RootState) => state.AccountReducer)
   const {isLogin} = loginState
   const dispatch = useDispatch();
