+++ src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import store from './store/store';
+import store from './redux/store';
 
 ReactDOM.render(
   <Provider store={store}>
