--- src/reducers/initialstate.tsx
@@ -1,8 +1,0 @@
-
-export type loginState = {
-  isLogin: boolean;
-};
-
-export const initialState: loginState = {
-  isLogin: false
-};
