--- src/reducers/loginReducer.tsx
@@ -1,18 +1,0 @@
-import {LOGIN, LOGOUT, LoginAction} from '../actions/actions';
-import { initialState, loginState} from './initialstate';
-import { createReducer } from 'typesafe-actions';
-
-const LoginReducer = createReducer<loginState, LoginAction>(initialState, {
-  [LOGIN]: (state, action) => {
-    return Object.assign({}, state, {
-      'isLogin': action.payload
-    })
-  },
-  [LOGOUT]: (state, action) => {
-    return Object.assign({}, state, {
-      'isLogin': action.payload
-    })
-  }
-})
-
-export default LoginReducer;
