--- src/store/store.tsx
@@ -1,9 +1,0 @@
-
-import {createStore } from "redux";
-import rootReducer  from '../reducers/index';
-
-
-const store = createStore(rootReducer);
-
-export default store;
-
